requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-core==2.4.1
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.26.3
pandas==2.1.4
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
totoapicontroller==1.0.0
tqdm==4.66.1
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
